backend/requirements.txt
annotated-types==0.7.0
autogen-agentchat==0.6.1
autogen-core==0.6.1
blinker==1.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.4
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonref==1.1.0
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.3.1
opentelemetry-api==1.34.1
packaging==25.0
pandas==2.3.0
pillow==11.2.1
protobuf==5.29.5
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
seaborn==0.13.2
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
